Update README.md
`rails s` by itself does not grant local network access. This is a problem if you are not using iOS simulator for the react native apps. There does seem to be a way to configure rails to use those parameters by default, but this works with the current set up.